rename art directory. remove dupe canvas
