+++ src/features/cosoul/art/lib/quaternion.js
