--- src/ui/Box/Canvas.tsx
@@ -1,5 +1,0 @@
-import { styled } from 'stitches.config';
-
-export const Canvas = styled('canvas', {});
-
-export default Canvas;
